# backend/src/main/java/com/volunteer/participation/controller/RequestsApiController.java
package com.volunteer.participation.controller;

import com.volunteer.common.dto.Error;
import com.volunteer.event.dto.CreateIndividualEventRequest;
import com.volunteer.event.dto.CreateMassEventRequest;
import com.volunteer.event.service.EventService;
import com.volunteer.entity.IndividualEvent;
import com.volunteer.entity.MassEvent;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import java.time.LocalDate;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@Controller
@RequestMapping("${openapi.volunteerInformationSystem.base-path:/api/v1}")
public class RequestsApiController implements RequestsApi {

    private final NativeWebRequest request;
    private final EventService eventService;

    @Autowired
    public RequestsApiController(NativeWebRequest request, EventService eventService) {
        this.request = request;
        this.eventService = eventService;
    }

    @Override
    public Optional<NativeWebRequest> getRequest() {
        return Optional.ofNullable(request);
    }

    // Mass Events

    @GetMapping("/events/mass")
    public ResponseEntity<List<MassEvent>> eventsMassGet(
            @RequestParam(required = false) String title,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateFrom,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateTo,
            @RequestParam(required = false) String address,
            @RequestParam(required = false) Integer organisationId) {
        try {
            List<MassEvent> events = eventService.getMassEvents(title, dateFrom, dateTo, address, organisationId);
            return ResponseEntity.ok(events);
        } catch (Exception e) {
            // Log error for debugging
            e.printStackTrace();
            // Return empty list on error instead of 500
            return ResponseEntity.ok(new java.util.ArrayList<>());
        }
    }

    @PostMapping("/events/mass")
    public ResponseEntity<MassEvent> eventsMassPost(@Valid @RequestBody CreateMassEventRequest createMassEventRequest) {
        try {
            MassEvent event = eventService.createMassEvent(createMassEventRequest);
            return ResponseEntity.status(HttpStatus.CREATED).body(event);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("organization representatives")) {
                throw new RuntimeException("FORBIDDEN: " + e.getMessage());
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @GetMapping("/events/mass/{id}")
    public ResponseEntity<MassEvent> eventsMassIdGet(@PathVariable("id") @NotNull Integer id) {
        try {
            MassEvent event = eventService.getMassEvent(id);
            return ResponseEntity.ok(event);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @PutMapping("/events/mass/{id}")
    public ResponseEntity<MassEvent> eventsMassIdPut(
            @PathVariable("id") @NotNull Integer id,
            @Valid @RequestBody CreateMassEventRequest createMassEventRequest) {
        try {
            MassEvent event = eventService.updateMassEvent(id, createMassEventRequest);
            return ResponseEntity.ok(event);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            if (e.getMessage().contains("organization representatives") || e.getMessage().contains("organization")) {
                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @GetMapping("/events/mass/{id}/parameters")
    public ResponseEntity<Map<String, String>> eventsMassIdParametersGet(@PathVariable("id") @NotNull Integer id) {
        try {
            // Verify event exists
            eventService.getMassEvent(id);
            // Return empty map as per API spec - this endpoint exists but returns no additional parameters
            Map<String, String> parameters = new HashMap<>();
            return ResponseEntity.ok(parameters);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    // Individual Events

    @GetMapping("/events/individual")
    public ResponseEntity<List<IndividualEvent>> eventsIndividualGet(
            @RequestParam(required = false) String title,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateFrom,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dateTo,
            @RequestParam(required = false) Integer organisationId) {
        try {
            List<IndividualEvent> events = eventService.getIndividualEvents(title, dateFrom, dateTo, organisationId);
            return ResponseEntity.ok(events);
        } catch (Exception e) {
            // Log the error for debugging
            e.printStackTrace();
            // Return empty list on error instead of 500, as per API expectations
            return ResponseEntity.ok(new java.util.ArrayList<>());
        }
    }

    @PostMapping("/events/individual")
    public ResponseEntity<IndividualEvent> eventsIndividualPost(@Valid @RequestBody CreateIndividualEventRequest createIndividualEventRequest) {
        try {
            IndividualEvent event = eventService.createIndividualEvent(createIndividualEventRequest);
            return ResponseEntity.status(HttpStatus.CREATED).body(event);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @GetMapping("/events/individual/{id}")
    public ResponseEntity<IndividualEvent> eventsIndividualIdGet(@PathVariable("id") @NotNull Integer id) {
        try {
            IndividualEvent event = eventService.getIndividualEvent(id);
            return ResponseEntity.ok(event);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @PutMapping("/events/individual/{id}")
    public ResponseEntity<IndividualEvent> eventsIndividualIdPut(
            @PathVariable("id") @NotNull Integer id,
            @Valid @RequestBody CreateIndividualEventRequest createIndividualEventRequest) {
        try {
            IndividualEvent event = eventService.updateIndividualEvent(id, createIndividualEventRequest);
            return ResponseEntity.ok(event);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            if (e.getMessage().contains("creator") || e.getMessage().contains("author")) {
                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    @GetMapping("/events/individual/{id}/parameters")
    public ResponseEntity<Map<String, String>> eventsIndividualIdParametersGet(@PathVariable("id") @NotNull Integer id) {
        try {
            // Verify event exists
            eventService.getIndividualEvent(id);
            // Return empty map as per API spec - this endpoint exists but returns no additional parameters
            Map<String, String> parameters = new HashMap<>();
            return ResponseEntity.ok(parameters);
        } catch (RuntimeException e) {
            if (e.getMessage().contains("not found")) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
            }
            throw new RuntimeException(e);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}