# doc/part-3/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Volunteer Information System API
  description: API для системы управления волонтерскими мероприятиями
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Чтение данных
            write: Запись данных
  schemas:
...
    MassEvent:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор массового мероприятия
        organisationId:
          type: integer
          description: ID организации-организатора
        title:
          type: string
          description: Название мероприятия
        description:
          type: string
          description: Описание мероприятия
        dateStart:
          type: string
          format: date-time
          description: Дата и время начала
        dateEnd:
          type: string
          format: date-time
          description: Дата и время окончания
        workHours:
          type: number
          format: float
          description: Количество часов работы для волонтера
        address:
          type: string
          description: Адрес проведения мероприятия
        volunteersRequired:
          type: integer
          description: Требуемое количество волонтеров
        ageRestriction:
          type: integer
          nullable: true
          description: Возрастное ограничение
    IndividualEvent:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор индивидуальной заявки
        organisationId:
          type: integer
          nullable: true
          description: ID организации (если заявка от организации)
        title:
          type: string
          description: Название заявки
        description:
          type: string
          description: Описание заявки
        dateStart:
          type: string
          format: date-time
          description: Дата и время начала
        dateEnd:
          type: string
          format: date-time
          description: Дата и время окончания
        volunteersRequired:
          type: integer
          description: Требуемое количество волонтеров
        ageRestriction:
          type: integer
          nullable: true
          description: Возрастное ограничение
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    CreateMassEventRequest:
      type: object
      required:
        - title
        - description
        - dateStart
        - dateEnd
        - workHours
        - address
        - volunteersRequired
      properties:
        title:
          type: string
          description: Название мероприятия
        description:
          type: string
          description: Описание мероприятия
        dateStart:
          type: string
          format: date-time
          description: Дата и время начала
        dateEnd:
          type: string
          format: date-time
          description: Дата и время окончания
        workHours:
          type: number
          format: float
          description: Количество часов работы для волонтера
        address:
          type: string
          description: Адрес проведения мероприятия
        volunteersRequired:
          type: integer
          minimum: 1
          description: Требуемое количество волонтеров
        ageRestriction:
          type: integer
          nullable: true
          description: Возрастное ограничение
    CreateIndividualEventRequest:
      type: object
      required:
        - title
        - description
        - dateStart
        - dateEnd
        - volunteersRequired
      properties:
        title:
          type: string
          description: Название заявки
        description:
          type: string
          description: Описание заявки
        dateStart:
          type: string
          format: date-time
          description: Дата и время начала
        dateEnd:
          type: string
          format: date-time
          description: Дата и время окончания
        volunteersRequired:
          type: integer
          minimum: 1
          description: Требуемое количество волонтеров
        ageRestriction:
          type: integer
          nullable: true
          description: Возрастное ограничение
...
paths:
...
  /events/mass:
    get:
      tags: [requests]
      summary: Получение списка массовых мероприятий
      description: Возвращает список всех массовых волонтерских мероприятий с возможностью фильтрации по параметрам
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Поиск по названию мероприятия (частичное совпадение)
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Поиск мероприятий начинающихся с указанной даты
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Поиск мероприятий заканчивающихся до указанной даты
        - name: address
          in: query
          required: false
          schema:
            type: string
          description: Поиск по адресу проведения (частичное совпадение)
        - name: organisationId
          in: query
          required: false
          schema:
            type: integer
          description: Поиск по ID организации-организатора
      responses:
        '200':
          description: Список массовых мероприятий успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MassEvent'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [requests]
      summary: Создание массового мероприятия
      description: Представитель организации создает новое массовое волонтерское мероприятие от имени своей организации
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMassEventRequest'
      responses:
        '201':
          description: Мероприятие успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MassEvent'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (требуется роль ORG_REPRESENTATIVE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Individual Events
  /events/individual:
    get:
      tags: [requests]
      summary: Получение списка индивидуальных заявок
      description: Возвращает список всех индивидуальных заявок с возможностью фильтрации по параметрам
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Поиск по названию заявки (частичное совпадение)
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Поиск заявок начинающихся с указанной даты
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Поиск заявок заканчивающихся до указанной даты
        - name: organisationId
          in: query
          required: false
          schema:
            type: integer
          description: Поиск по ID организации (если заявка от организации)
      responses:
        '200':
          description: Список индивидуальных заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndividualEvent'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [requests]
      summary: Создание индивидуальной заявки
      description: Авторизованный пользователь создает новую индивидуальную волонтерскую заявку
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndividualEventRequest'
      responses:
        '201':
          description: Заявка успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualEvent'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/mass/{id}:
    get:
      tags: [requests]
      summary: Получение деталей массового мероприятия
      description: Возвращает детальную информацию о массовом мероприятии по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '200':
          description: Детали мероприятия успешно получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MassEvent'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [admin]
      summary: Удаление массового мероприятия
      description: Администратор удаляет массовое мероприятие в целях модерации
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '204':
          description: Мероприятие успешно удалено
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [requests]
      summary: Редактирование массового мероприятия
      description: Представитель организации редактирует массовое мероприятие своей организации
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMassEventRequest'
      responses:
        '200':
          description: Мероприятие успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MassEvent'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (требуется роль ORG_REPRESENTATIVE и принадлежность к организации)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{id}:
    get:
      tags: [requests]
      summary: Получение деталей индивидуальной заявки
      description: Возвращает детальную информацию об индивидуальной заявке по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '200':
          description: Детали заявки успешно получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualEvent'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [admin]
      summary: Удаление индивидуальной заявки
      description: Администратор удаляет индивидуальную заявку в целях модерации
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '204':
          description: Заявка успешно удалена
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [requests]
      summary: Редактирование индивидуальной заявки
      description: Автор заявки редактирует свою индивидуальную заявку
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndividualEventRequest'
      responses:
        '200':
          description: Заявка успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualEvent'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (пользователь не является автором заявки)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /events/mass/{id}/parameters:
    get:
      tags: [requests]
      summary: Получение параметров массового мероприятия
      description: Возвращает дополнительные параметры массового мероприятия по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '200':
          description: Параметры мероприятия успешно получены
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{id}/parameters:
    get:
      tags: [requests]
      summary: Получение параметров индивидуальной заявки
      description: Возвращает дополнительные параметры индивидуальной заявки по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '200':
          description: Параметры заявки успешно получены
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Participants Management
...